Added Map Struct
