--- src/main/java/com/gamedoora/model/dto/GdStory.java
@@ -1,109 +1,0 @@
-package com.gamedoora.model.dto;
-
-
-import java.util.Date;
-
-public class GdStory {
-
-  private long id;
-  private String name;
-  private String description;
-  private long startDate;
-  private long dueDate;
-  private long stateId;
-  private long position;
-  private String remarks;
-  private Date createdAt;
-  private Date updatedAt;
-
-
-  public long getId() {
-    return id;
-  }
-
-  public void setId(long id) {
-    this.id = id;
-  }
-
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-
-  public String getDescription() {
-    return description;
-  }
-
-  public void setDescription(String description) {
-    this.description = description;
-  }
-
-
-  public long getStartDate() {
-    return startDate;
-  }
-
-  public void setStartDate(long startDate) {
-    this.startDate = startDate;
-  }
-
-
-  public long getDueDate() {
-    return dueDate;
-  }
-
-  public void setDueDate(long dueDate) {
-    this.dueDate = dueDate;
-  }
-
-
-  public long getStateId() {
-    return stateId;
-  }
-
-  public void setStateId(long stateId) {
-    this.stateId = stateId;
-  }
-
-
-  public long getPosition() {
-    return position;
-  }
-
-  public void setPosition(long position) {
-    this.position = position;
-  }
-
-
-  public String getRemarks() {
-    return remarks;
-  }
-
-  public void setRemarks(String remarks) {
-    this.remarks = remarks;
-  }
-
-
-  public Date getCreatedAt() {
-    return createdAt;
-  }
-
-  public void setCreatedAt(Date createdAt) {
-    this.createdAt = createdAt;
-  }
-
-
-  public Date getUpdatedAt() {
-    return updatedAt;
-  }
-
-  public void setUpdatedAt(Date updatedAt) {
-    this.updatedAt = updatedAt;
-  }
-
-}
