--- src/main/java/com/gamedoora/model/dto/GdTenant.java
@@ -1,69 +1,0 @@
-package com.gamedoora.model.dto;
-
-
-import java.util.Date;
-
-public class GdTenant {
-
-  private long id;
-  private String subdomain;
-  private long entityId;
-  private String entityType;
-  private Date createdAt;
-  private Date updatedAt;
-
-
-  public long getId() {
-    return id;
-  }
-
-  public void setId(long id) {
-    this.id = id;
-  }
-
-
-  public String getSubdomain() {
-    return subdomain;
-  }
-
-  public void setSubdomain(String subdomain) {
-    this.subdomain = subdomain;
-  }
-
-
-  public long getEntityId() {
-    return entityId;
-  }
-
-  public void setEntityId(long entityId) {
-    this.entityId = entityId;
-  }
-
-
-  public String getEntityType() {
-    return entityType;
-  }
-
-  public void setEntityType(String entityType) {
-    this.entityType = entityType;
-  }
-
-
-  public Date getCreatedAt() {
-    return createdAt;
-  }
-
-  public void setCreatedAt(Date createdAt) {
-    this.createdAt = createdAt;
-  }
-
-
-  public Date getUpdatedAt() {
-    return updatedAt;
-  }
-
-  public void setUpdatedAt(Date updatedAt) {
-    this.updatedAt = updatedAt;
-  }
-
-}
