+++ src/main/java/com/gamedoora/model/mapper/CommentMapper.java
@@ -1,0 +1,14 @@
+package com.gamedoora.model.mapper;
+
+import org.mapstruct.Mapper;
+
+import com.gamedoora.model.dao.GdComment;
+import com.gamedoora.model.dto.Comment;
+
+@Mapper
+public interface CommentMapper {
+
+	Comment modelToDto(GdComment comment);
+
+	GdComment dtoToModel(Comment comment);
+}
